# Conflicts:
#	.gitignore
#	src/core/client/client.c
#	src/core/client/context.c
#	src/core/client/send.c
#	src/core/util/bytes.c
#	src/core/util/data.c
#	src/zephyr-motor-boot/CMakeLists.txt
#	src/zephyr-motor-boot/src/fsm.c
#	src/zephyr-motor-boot/src/main.c
#	src/zephyr-motor-boot/src/rent.c
#	src/zephyr/CMakeLists.txt
#	src/zephyr/prj.conf
#	src/zephyr/src/fsm.c
#	src/zephyr/src/fsm.h
#	src/zephyr/src/main.c
#	src/zephyr/src/rent.c

+++ src/eth_nano/txreceipt.c
@@ -122,7 +122,7 @@ int eth_verify_eth_getTransactionReceipt(in3_vctx_t* vc, bytes_t* tx_hash) {
         return vc_err(vc, "wrong block number in log");
       if (!d_eq(block_hash, d_get(l, K_BLOCK_HASH)))
         return vc_err(vc, "wrong block hash in log");
-      if (d_get_intk(l, K_LOG_INDEX) != i)
+      if (d_get_intk(l, K_LOG_INDEX) != (uint32_t) i)
         return vc_err(vc, "wrong log index");
       if (!b_cmp(d_get_bytesk(l, K_TRANSACTION_HASH), tx_hash))
         return vc_err(vc, "wrong tx Hash");
